tech(api): migrate cleanup-localized-challenges-and-translations script
